feat: define the hasSymbol function
